Use p5 random, duplicate functions where necessary
